+++ tests/utils/blockbinary_helpers.hpp
@@ -46,7 +46,7 @@ void ReportBinaryArithmeticSuccess(std::string test_case, std::string op, const 
 }
 
 template<size_t nbits, typename Ty = uint8_t>
-void ReportArithmeticShiftError(std::string test_case, std::string op, const sw::unum::blockbinary<nbits, Ty>& a, const long divider, const sw::unum::blockbinary<nbits, Ty>& result, int64_t reference) {
+void ReportArithmeticShiftError(std::string test_case, std::string op, const sw::unum::blockbinary<nbits, Ty>& a, const size_t divider, const sw::unum::blockbinary<nbits, Ty>& result, int64_t reference) {
 	using namespace sw::unum;
 	auto old_precision = std::cerr.precision();
 	std::cerr << test_case << " "
@@ -64,7 +64,7 @@ void ReportArithmeticShiftError(std::string test_case, std::string op, const sw:
 }
 
 template<size_t nbits, typename Ty = uint8_t>
-void ReportArithmeticShiftSuccess(std::string test_case, std::string op, const sw::unum::blockbinary<nbits, Ty>& a, const long divider, const sw::unum::blockbinary<nbits, Ty>& result, int64_t reference) {
+void ReportArithmeticShiftSuccess(std::string test_case, std::string op, const sw::unum::blockbinary<nbits, Ty>& a, const size_t divider, const sw::unum::blockbinary<nbits, Ty>& result, int64_t reference) {
 	using namespace sw::unum;
 	auto old_precision = std::cerr.precision();
 	std::cerr << test_case << " "
